requirements.txt
aiofiles
fastapi
openai-whisper
uvicorn[standard]